refactor(engine): move InterPartitionCommandSender to the common api between engine and streamprocessor
It is better that everything that is needed by the engine and managed by the system is defined via the common api. This way the api between the system and the engine will be clearer.
