+++ engine/src/main/java/io/camunda/zeebe/engine/api/RecordProcessorContext.java
@@ -11,7 +11,6 @@ import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDb;
 import io.camunda.zeebe.engine.state.EventApplier;
 import io.camunda.zeebe.engine.state.mutable.MutableZeebeState;
-import io.camunda.zeebe.engine.transport.InterPartitionCommandSender;
 import java.util.List;
 import java.util.function.Function;
